fixup! feat(stat/incr/nanmcovariance): add extra nanmcovariance test and benchmark cases
This commit fixes a coverage issue with the previous commit where a branch of the main.js file of the nanmcovariance package, is not adequetly tested.
It now ensures that the branch for known and unknown means are tested and benchmarked, for full coverage.

Authored by: Don Chacko <donisepic30@gmail.com>