feat: add `symbol/split`
This commit adds the `@stdlib/symbol/split` package which exports
`Symbol.split` if the environment supports it, otherwise `null`.

The `Symbol.split` symbol specifies a method that splits a string at
the indices that match a regular expression. When calling
`String.prototype.split` and the separator argument is an object
with a `[Symbol.split]()` method, this method is called with the
target string and limit as arguments.

The package structure follows the existing convention established
by similar packages such as `symbol/replace`, `symbol/has-instance`,
and `symbol/is-concat-spreadable`.

Closes: #8488